requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.5.18.1
cffi==1.15.1
chardet==3.0.4
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.0
defusedxml==0.7.1
Django==3.1.6
django-allauth==0.51.0
django-appconf==1.0.4
django-celery-beat==2.4.0
django-compressor==2.4
django-extensions==3.1.5
django-js-asset==1.2.2
django-redis==5.2.0
django-rest-auth==0.9.5
django-storages==1.9.1
django-timezone-field==5.0
django-user-agents==0.4.0
djangorestframework==3.11.0
docutils==0.15.2
drf-nested-routers==0.91
flower==1.2.0
future==0.18.2
greenlet==2.0.2
gunicorn==20.1.0
humanize==4.6.0
idna==2.10
jmespath==1.0.0
kombu==5.2.4
oauthlib==3.2.0
prompt-toolkit==3.0.37
psycopg2-binary==2.8.6
pycparser==2.21
PyJWT==2.4.0
python-crontab==2.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
rabbitmq==0.2.0
rcssmin==1.0.6
redis==4.5.1
requests==2.23.0
requests-oauthlib==1.3.0
rjsmin==1.1.0
s3transfer==0.5.2
serpy==0.3.1
six==1.16.0
SQLAlchemy==2.0.4
sqlparse==0.4.2
tornado==6.2
typing_extensions==4.5.0
tzdata==2022.7
ua-parser==0.10.0
urllib3==1.25.11
user-agents==2.2.0
vine==5.0.0
wcwidth==0.2.6
wget==3.2
